ui: Remove NewEventTimelineItem
